Move API types to declaration file
