+++ types/api.d.ts
@@ -1,3 +1,7 @@
+import { IEvidenceRequest } from 'src/domain/evidence-request';
+import { IDocumentType } from 'src/domain/document-type';
+import { IDocumentSubmission } from 'src/domain/document-submission';
+
 export type TokenDictionary = {
   [key: string]:
     | {
@@ -5,3 +9,22 @@ export type TokenDictionary = {
       }
     | undefined;
 };
+
+export type DeliveryMethodResponse = 'sms' | 'email';
+
+export interface EvidenceRequestResponse
+  extends Omit<
+    IEvidenceRequest,
+    'createdAt' | 'deliveryMethods' | 'documentTypes'
+  > {
+  createdAt: string;
+  deliveryMethods: string[];
+  documentTypes: IDocumentType[];
+}
+
+export interface DocumentSubmissionResponse
+  extends Omit<IDocumentSubmission, 'createdAt' | 'state' | 'documentType'> {
+  createdAt: string;
+  state: string;
+  documentType: IDocumentType;
+}
